==> 1_data_consolidation.ipynb <==
import pandas as pd
import numpy as np

import sqlite3

## Load in ODP dataset

odpcontentpath = "webdata/odp/content.rdf.u8"

PAGELEAD = "  <ExternalPage"
LINKPOS = len(PAGELEAD) + len(" about=")

TITLELEAD = "    <d:Title>"
DESCLEAD = "    <d:Description>"
TOPICLEAD = "    <topic>Top/"

ENDPAGE = "  </ExternalPage"


data =[]

link = ""
revdomain = ""
title = ""
description = ""
topic = ""

foundpage = False
n = 0
with open(odpcontentpath, "r", encoding="utf-8") as file:
    for line in file:
        if foundpage:
            if line.startswith(TITLELEAD):
                title = line[len(TITLELEAD):line.find("</")]
            elif line.startswith(DESCLEAD):
                description = line[len(DESCLEAD):line.find("</")]
            elif line.startswith(TOPICLEAD):
                topic = line[len(TOPICLEAD):line.find("</")]
                # remove the subtopics
                slash = topic.find("/")
                if slash != -1:
                    topic = topic[:slash]
            elif line.startswith(ENDPAGE):
                data.append([revdomain, link, title, description, topic])
                foundpage = False
                
        if line.startswith(PAGELEAD):
            foundpage = True
            n += 1
            link = line[LINKPOS+1:line.find(">")-1]
    
            domain = link.split("://")[-1].split("/")[0]
            if domain.startswith("www."):
                domain = domain[4:]
            revdomain = ".".join(reversed(domain.split(".")))
            if revdomain[0] == ".":
                revdomain = revdomain[1:]

odp_df = pd.DataFrame(data, columns = ["ReversedDomain", "URL", "Title", "Description", "Topic"])
odp_df.head(20)

odp_df.to_csv("outputs/odp_df.csv", index=False)

# odp_df = pd.read_csv("odp_df.csv")

len(odp_df)

## Load in Common Crawl web graph

verticiesdomainpath = "webdata/commoncrawl/cc-main-2017-may-jun-jul/vertices.txt"

ccdomains_df = pd.read_csv(
        verticiesdomainpath,
        sep="\t",
        names=["ID", "ReversedDomain"],  # Explicit column names
        dtype={"ID": "int32", "ReversedDomain": "string"},  # Use optimized data types
        engine="c",  # Use the C engine for faster parsing
    )

len(ccdomains_df)

ccdomains_df.head(20)

## Match the dataframes

conn = sqlite3.connect("outputs/matching.db")

odp_df.to_sql("odp", conn, if_exists="replace", index=False)
ccdomains_df.to_sql("ccdomains", conn, if_exists="replace", index=False)

conn.execute("CREATE INDEX idx_odp ON odp (ReversedDomain);")

conn.execute("CREATE INDEX idx_ccdomains ON ccdomains (ReversedDomain);")

pd.read_sql_query("SELECT ReversedDomain FROM odp;", conn)

query = """
    CREATE TABLE matched AS
    SELECT *
    FROM odp
    INNER JOIN ccdomains
    ON odp.ReversedDomain = ccdomains.ReversedDomain;
"""

conn.execute(query)

matched_df = pd.read_sql_query("SELECT * FROM matched", conn)

matched_df = matched_df.drop(['ReversedDomain:1'], axis=1)

matched_df

matched_df.to_csv("outputs/matched_df.csv", index=False)
