==> 2_vertex_cleaning.ipynb <==
import pandas as pd
import numpy as np
import re

matched_df = pd.read_csv("outputs/matched_df.csv")
matched_df

matched_df = matched_df.replace(np.nan, "")

matched_df = matched_df.drop_duplicates()
len(matched_df)

# create a content column 
matched_df["AllContent"] = matched_df["Title"] + " " + matched_df["Description"]

# Create a cleaned content column

double_dash_regex = re.compile(r"--+")  # Matches double dashes
invalid_char_regex = re.compile(r"[^\w\s-]+")  # Matches non-alphanumeric, non-space, non-dash characters
underscore_regex = re.compile(r"_")  # Matches underscores
numbers_regex = re.compile(r"\b[\w-]*\d[\w-]*\b")  # Matches valid words without digits
leading_trailing_dash_regex = re.compile(r"(?<!\w)-+|-+(?!\w)")  # Matches leading/trailing dashes

# Replace double dashes with a space
matched_df["CleanedContent"] = (
    matched_df["AllContent"]
    .str.replace(double_dash_regex, " ", regex=True)  # Replace double dashes 
    .str.replace(invalid_char_regex, " ", regex=True) # Replace invalid characters with a space (non-ASCII, non-alphanumeric, and non-dash)
    .str.replace(underscore_regex, " ", regex=True)  # Replace underscores with spaces
    .str.replace(numbers_regex, "", regex=True)  # Remove words with numbers
    .str.replace(leading_trailing_dash_regex, "", regex=True) # Strip leading and trailing dashes for each word
    .str.lower()
)

matched_df.head(20)

len(matched_df)

# Keep only ASCII words
matched_df["CleanedContent"] = matched_df["CleanedContent"].str.split().apply(lambda words: " ".join([word for word in words if word.isascii()]))

# Drop rows with empty CleanedContents
matched_df = matched_df[matched_df["CleanedContent"].str.len() > 0].reset_index(drop=True)

matched_df.head(20)

len(matched_df)

# drop duplicates where the only difference is the description (keep first version)
matched_df = matched_df.drop_duplicates(subset=['URL', 'Title', 'Topic', 'ID'], keep="first")
len(matched_df)

# if a reverseddomain appears more than once and it is not the case that only the topic and description differs
# then drop all instances of that reversedomain from the dataframe
# (ie. drop all rows unless the only difference is the topic or description)
# WE DO THIS IN ORDER TO PREVENT PAGES WITH LITTLE RELEVANCE TO BE PRIORITIZED DUE TO THEM BELONGING TO A 
# DOMAIN WITH A HIGH NUMBER OF INBOUND LINKS

columns_to_check = matched_df.columns.difference(['ReversedDomain', 'Description', 'Topic', 'AllContent', 'CleanedContent'])
non_topic_variation = matched_df.groupby('ReversedDomain')[columns_to_check].nunique().max(axis=1)
domains_to_remove = non_topic_variation[non_topic_variation > 1].index
domains_to_remove

matched_df = matched_df[~matched_df['ReversedDomain'].isin(domains_to_remove)].reset_index(drop=True)
matched_df

####

THEN WE WANT TO USE THIS CLEANED VERSION WITH THE SEPARATE CATAGEOIRES 
BUT ALSO CREATE A "PHONEBOOK" WITH THE FIRST REPEAT OF THE DOMAIN TAKEN

###

# WE CREATE A "PHONEBOOK" WITH THE FIRST REPEAT OF THE DOMAIN TAKEN
pagelookup_df = matched_df[["ReversedDomain", "URL", "Title", "Description", "ID"]].drop_duplicates(
    subset=['ReversedDomain'], keep='first').reset_index(drop=True)
pagelookup_df

pagelookup_df.to_csv('outputs/pagelookup_df.csv', index=False)

cleaned_df = matched_df[["ReversedDomain", "CleanedContent", "Topic", "ID"]]
cleaned_df = cleaned_df.rename(columns={"CleanedContent": "Content"})
cleaned_df.head(20)

len(cleaned_df)

cleaned_df.to_csv('outputs/cleaned_matched_df.csv', index=False)

## Create list of all terms

terms = cleaned_df["Content"].str.split().explode().drop_duplicates()
terms

len(terms)

terms_df = pd.DataFrame({'Terms': terms})

terms_df = terms_df.reset_index(drop=True)
terms_df

## Create term-vectors

exploded_df = cleaned_df.copy()
exploded_df["Content"] = exploded_df["Content"].str.split()
exploded_df = exploded_df.explode("Content")

term_counts = (
    exploded_df.groupby("Topic")["Content"]
    .value_counts()
    .reset_index(name="Occurrences")
)
term_counts

term_counts = term_counts.pivot(index="Content", columns="Topic", values="Occurrences").fillna(0).astype(int)
term_counts.reset_index(inplace=True)
term_counts

term_counts = term_counts.rename(columns={"Content": "Terms"})

term_vectors_df = terms_df.merge(
    term_counts,  # Existing term counts
    on="Terms",  # Match on the 'Content' column
    how="left"  # Ensure all terms in all_terms_df are kept
)
term_vectors_df

term_vectors_df.to_csv('outputs/term_vectors.csv', index=False)

# Results

cleaned_df

term_vectors_df
