==> 3_rank_vectors_creation.ipynb <==
import pandas as pd
import numpy as np
from scipy import sparse

from fast_pagerank import pagerank_power

cleaned_df = pd.read_csv("outputs/cleaned_matched_df.csv")
cleaned_df

edges_df = pd.read_csv("outputs/filtered_edges_df.csv")
edges_df

valid_nodes = cleaned_df["ID"].drop_duplicates().reset_index(drop=True)
valid_nodes

## Create graph

# Create indexing for graph (convert ID to IND)
id_node_series = pd.Series(valid_nodes.index, index=valid_nodes.values)
id_node_series

dim = len(valid_nodes)
G = sparse.csr_matrix((np.ones(len(edges_df)), 
                       (id_node_series.reindex(edges_df['from_node']), 
                        id_node_series.reindex(edges_df['to_node']))), shape=(dim, dim))

## Create topic biased pagerank vectors (+ personalization vectors)

pr_vectors = valid_nodes.to_frame()
personal_vectors = valid_nodes.to_frame()
personal_vectors

alpha = 0.85

for topic, group in cleaned_df.groupby("Topic"):
    personalization_vector = np.zeros(dim, dtype=float)
    personalization_vector[id_node_series.reindex(group["ID"])] = 1 / len(group)
    
    personal_vectors[topic] = personalization_vector

    pr_vectors[topic] = pagerank_power(G, p=alpha, personalize=personalization_vector, tol=1e-9)

personal_vectors

pr_vectors

## Create unbiased pagerank vector

pr_vectors["Unbiased"] = pagerank_power(G, p=alpha, tol=1e-9)

pr_vectors

pr_vectors.to_csv("outputs/pr_vectors_df.csv", index=False)

personal_vectors.to_csv("outputs/personalization_vectors_df.csv", index=False)
